refact(step4): restructure packages
