--- src/main/java/ladder/entity/MoveRight.java
@@ -1,9 +1,0 @@
-package ladder.entity;
-
-public class MoveRight implements MoveStrategy{
-
-    @Override
-    public Direction move() {
-        return Direction.RIGHT;
-    }
-}
