--- src/main/java/ladder/entity/MoveStrategy.java
@@ -1,5 +1,0 @@
-package ladder.entity;
-
-public interface MoveStrategy {
-    Direction move();
-}
