+++ src/main/java/ladder/entity/ladderMap/LadderHeight.java
@@ -1,4 +1,4 @@
-package ladder.entity;
+package ladder.entity.ladderMap;
 
 import ladder.exception.CustomException;
 import ladder.exception.ErrorCode;
@@ -7,12 +7,12 @@ public class LadderHeight {
 
     private int height;
 
-    public LadderHeight(int height){
+    public LadderHeight(int height) {
         this.height = verifiedHeight(height);
     }
 
-    private int verifiedHeight(int height){
-        if(height<=0){
+    private int verifiedHeight(int height) {
+        if (height <= 0) {
             throw new CustomException(ErrorCode.INVALID_LADDER_HEIGHT);
         }
         return height;
