+++ src/main/java/ladder/entity/moveStrategy/MoveLeft.java
@@ -1,4 +1,6 @@
-package ladder.entity;
+package ladder.entity.moveStrategy;
+
+import ladder.entity.Direction;
 
 public class MoveLeft implements MoveStrategy {
