+++ src/main/java/ladder/entity/moveStrategy/MoveRight.java
@@ -1,0 +1,11 @@
+package ladder.entity.moveStrategy;
+
+import ladder.entity.Direction;
+
+public class MoveRight implements MoveStrategy {
+
+    @Override
+    public Direction move() {
+        return Direction.RIGHT;
+    }
+}
