+++ src/main/java/ladder/entity/moveStrategy/MoveStill.java
@@ -1,4 +1,6 @@
-package ladder.entity;
+package ladder.entity.moveStrategy;
+
+import ladder.entity.Direction;
 
 public class MoveStill implements MoveStrategy {
