+++ src/main/java/ladder/entity/moveStrategy/MoveStrategy.java
@@ -1,0 +1,7 @@
+package ladder.entity.moveStrategy;
+
+import ladder.entity.Direction;
+
+public interface MoveStrategy {
+    Direction move();
+}
